Add test pattern image
This is based on the output of the Philips PM5544 television pattern
generator. The goal is to create a useful image to test LCD panels
though, not analog televisions, so I modified some details to be more
useful for finding the right settings for an LCD panel. It would
probably be useful to modify it more in the future.

The image is based on the SVG from Wikipedia user "ebnz". I re-aligned
all the elements to fit the 240x160 resolution of the RS90's LCD.